requirements.txt
yfinance
feedparser
python-dotenv
pinecone
requests
openai
fastapi